capstone update

+++ app/Http/Controllers/Backend/Inventory/SuppliersController.php
@@ -4,10 +4,12 @@ namespace App\Http\Controllers\Backend\Inventory;
 
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
-use App\Supplier;
+use App\Notifications\SupplierResource;
+use App\{Supplier,User};
 use Auth;
 use Hash;
 use Image;
+use DB;
 
 class SuppliersController extends Controller
 {
@@ -184,15 +186,44 @@ class SuppliersController extends Controller
     {
         if(Hash::check(request('password'),Auth::user()->password)){
             $supplier = Supplier::find(request('supplier_id'));
-            $supplier->delete();
+            $count = DB::table('products')
+                        ->where('supplier_id',$supplier->supplier_id)
+                        ->where('deleted_at','=',null)
+                        ->count();
 
-            $name = $supplier->getFullName();
-            $notification = array (
-                'message' => "Supplier ".$name." was successfuly deleted!",
-                'icon' => 'success',
-                'heading' => 'Success',
-            );
-            return back()->with($notification);
+            if($count == 0)
+            {
+                $auth_user  = Auth::user();
+                $admins = User::where('role_id',1)->get();
+                $badge = array(
+                    'bg' => "danger",
+                    'icon' => "fa fa-address-card mx-0",
+                );
+                $supplier->delete();
+                
+                foreach($admins as $admin)
+                {
+                    $admin->notify(new SupplierResource($auth_user,$supplier,"archived",$badge));
+                }
+
+                $name = $supplier->getFullName();
+                $notification = array (
+                    'message' => "Supplier ".$name." was successfuly deleted!",
+                    'icon' => 'success',
+                    'heading' => 'Success',
+                );
+                return back()->with($notification);
+            }
+            else
+            {
+                $notification = array(
+                    'message' => "Supplier is being used by other/s Product!",
+                    'icon'  => 'error',
+                    'heading'   => 'Failed!',
+                );
+                
+                return back()->with($notification);
+            }
         }
         else{
             $notification = array (
@@ -202,5 +233,40 @@ class SuppliersController extends Controller
             );
             return back()->with($notification);
         }
+    }
+
+    public function restore(Request $request)
+    {
+        $supplier = Supplier::onlyTrashed()
+                            ->where('supplier_id',request('supplier_id'))
+                            ->firstOrFail();
+        $auth_user = Auth::user();
+        $admins = User::where('role_id',1)->get();
+        $badge = array(
+            'bg' => "info",
+            'icon' => "fa fa-address-card mx-0",
+        );
+
+        $supplier->restore();
+
+        foreach($admins as $admin)
+        {
+            $admin->notify(new SupplierResource($auth_user,$supplier,"restored",$badge));
+        }
+       
+        $notification = array(
+            'message' => 'Supplier successfully Restored!',
+            'icon'    => 'success',
+            'heading' => 'SUccess!',
+        );
+
+        return back()->with($notification);
+    }
+
+    public function archive()
+    {
+        $suppliers = Supplier::onlyTrashed()->get();
+
+        return view('backend.admin.suppliers.suppliersArchive',compact('suppliers'));
     }
 }
